feat: support `scanOptions.template` option.
