+++ package.json
@@ -18,10 +18,10 @@
     "lint": "eslint . --fix --ext .js,.ts"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.57.0",
-    "@typescript-eslint/parser": "^5.57.0",
+    "@typescript-eslint/eslint-plugin": "^5.58.0",
+    "@typescript-eslint/parser": "^5.58.0",
     "@vue/eslint-config-typescript": "^11.0.2",
-    "eslint": "^8.37.0",
+    "eslint": "^8.38.0",
     "eslint-config-standard": "^17.0.0",
     "eslint-define-config": "^1.17.0",
     "eslint-plugin-import": "^2.27.5",
@@ -32,7 +32,7 @@
     "typescript": "^4.9.5",
     "vue-eslint-parser": "^9.1.1"
   },
-  "packageManager": "pnpm@7.14.0",
+  "packageManager": "pnpm@8.0.2",
   "pnpm": {
     "overrides": {
       "vite-plugin-virtual-mpa": "workspace:*"
