+++ workspaces/plugin/package.json
@@ -1,6 +1,6 @@
 {
   "name": "vite-plugin-virtual-mpa",
-  "version": "1.8.1",
+  "version": "1.8.2",
   "license": "MIT",
   "author": "秦旭洋",
   "main": "dist/index.js",
@@ -50,17 +50,17 @@
     "@types/connect-history-api-fallback": "^1.3.5",
     "@types/html-minifier-terser": "^7.0.0",
     "connect-history-api-fallback": "^2.0.0",
-    "ejs": "^3.1.8",
-    "html-minifier-terser": "^7.1.0",
+    "ejs": "^3.1.9",
+    "html-minifier-terser": "^7.2.0",
     "picocolors": "^1.0.0"
   },
   "devDependencies": {
-    "@types/ejs": "^3.1.1",
-    "@types/node": "^18.14.0",
-    "typescript": "^4.9.4",
+    "@types/ejs": "^3.1.2",
+    "@types/node": "^18.15.11",
+    "typescript": "^4.9.5",
     "vite": "^4.2.1",
     "vite-plugin-checker": "^0.5.6",
-    "vite-plugin-dts": "^2.1.0",
+    "vite-plugin-dts": "^2.2.0",
     "vite-plugin-externalize-deps": "^0.5.0"
   },
   "peerDependencies": {
