+++ workspaces/plugin/src/plugin.ts
@@ -4,8 +4,15 @@ import fs from 'fs';
 import path from 'path';
 import history, { Rewrite } from 'connect-history-api-fallback';
 import { name as pkgName } from '../package.json';
-import type { MpaOptions, AllowedEvent, Page, WatchOptions, ScanOptions } from './api-types';
-import { type ResolvedConfig, type Plugin, normalizePath, createFilter, ViteDevServer } from 'vite';
+import type { MpaOptions, AllowedEvent, Page, WatchOptions } from './api-types';
+import { scanPages } from './utils';
+import {
+  type ResolvedConfig,
+  type Plugin,
+  type ViteDevServer,
+  normalizePath,
+  createFilter,
+} from 'vite';
 
 const bodyInject = /<\/body>/;
 const pluginName = color.cyan(pkgName);
@@ -299,33 +306,6 @@ export function createMpaPlugin<
 
 function throwError(message) {
   throw new Error(`[${pluginName}]: ${color.red(message)}`);
-}
-
-/**
- * Generate pages configurations using scanOptions.
- */
-function scanPages(scanOptions?: ScanOptions) {
-  const { filename, entryFile, scanDirs } = scanOptions || {} as ScanOptions;
-  const pages: Page[] = [];
-
-  for (const entryDir of [scanDirs].flat().filter(Boolean)) {
-    for (const name of fs.readdirSync(entryDir)) {
-      const dir = path.join(entryDir, name); // dir path
-      if (!fs.statSync(dir).isDirectory()) continue;
-
-      pages.push({
-        name,
-        filename: typeof filename === 'function'
-          ? filename(name) as Page['filename']
-          : undefined,
-        entry: entryFile
-          ? path.join('/', dir, entryFile) as Page['entry']
-          : undefined,
-      });
-    }
-  }
-
-  return pages;
 }
 
 // // This is for type declaration testing.
